mount_utils: move mount_at() and mount_from_at() over from utils.{c,h}
