+++ src/lxc/mount_utils.h
@@ -3,6 +3,7 @@
 #ifndef __LXC_MOUNT_UTILS_H
 #define __LXC_MOUNT_UTILS_H
 
+#include <linux/types.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <sys/mount.h>
@@ -149,5 +150,17 @@ __hidden extern int mnt_attributes_new(unsigned int old_flags, unsigned int *new
 __hidden extern int mnt_attributes_old(unsigned int new_flags, unsigned int *old_flags);
 
 __hidden extern int mount_filesystem(const char *fs_name, const char *path, unsigned int attr_flags);
+
+__hidden extern int mount_at(int dfd, const char *src_under_dfd,
+			     const char *dst_under_dfd, __u64 o_flags,
+			     __u64 resolve_flags, const char *fstype,
+			     unsigned int mnt_flags, const void *data);
+
+__hidden extern int mount_from_at(int dfd_from, const char *path_from,
+				  __u64 o_flags_from, __u64 resolve_flags_from,
+				  int dfd_to, const char *path_to,
+				  __u64 o_flags_to, __u64 resolve_flags_to,
+				  const char *fstype, unsigned int mnt_flags,
+				  const void *data);
 
 #endif /* __LXC_MOUNT_UTILS_H */
